chore: move `binary_or_hex_heuristic` under API, and proptest it
