+++ .gitignore
@@ -3,6 +3,8 @@
 debug/
 target/
 
+proptest-regressions/
+
 # These are backup files generated by rustfmt
 **/*.rs.bk
