+++ Cargo.toml
@@ -49,6 +49,7 @@ anyhow = "1.0.95"
 tracing-test = "0.2.5"
 rstest = "0.24.0"
 pretty_assertions = "1.4.1"
+proptest = "1.6"
 
 [target.'cfg(target_env = "musl")'.dependencies]
 jemalloc = "0.3"
